# bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DealerPro",
    "version": "v1"
  },
  "paths": {
    "/api/addCustomProducts/addProduct": {
      "post": {
        "tags": [
          "addCustomProducts"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "UID",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "StoreName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Name",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Img",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "Category",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Image",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Price",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "OldPrice",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Brand",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "description",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "in": "formData",
            "name": "ProductUrl",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Date",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Time",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/addCustomProducts/updateProduct/{category}/{productId}": {
      "put": {
        "tags": [
          "addCustomProducts"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "category",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "StoreName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Name",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Img",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "Image",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Price",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "OldPrice",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Brand",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "ProductUrl",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "description",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "in": "formData",
            "name": "Date",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Time",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/addCustomProducts/deleteProduct/{category}/{productId}": {
      "delete": {
        "tags": [
          "addCustomProducts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "category",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
}